+++ babel-preset/configs/hot.js
