+++ babel-preset/configs/internal.js
