+++ packager/react-packager/src/JSTransformer/worker.js
@@ -9,27 +9,25 @@
 'use strict';
 
 var babel = require('babel-core');
-var makeInternalPreset = require('babel-preset-react-native/internal');
+var makeInternalConfig = require('babel-preset-react-native/configs/internal');
 
 // Runs internal transforms on the given sourceCode. Note that internal
 // transforms should be run after the external ones to ensure that they run on
 // Javascript code
 function internalTransforms(sourceCode, filename, options) {
-  var internalPreset = makeInternalPreset(options);
+  var internalBabelConfig = makeInternalConfig(options);
 
-  if (!internalPreset) {
+  if (!internalBabelConfig) {
     return {
       code: sourceCode,
       filename: filename,
     };
   }
 
-  var presets = [ internalPreset ];
-
   var result = babel.transform(sourceCode, {
     filename: filename,
     sourceFileName: filename,
-    presets: presets
+    ...internalBabelConfig
   });
 
   return {
